agent/requirements.txt
fastapi
uvicorn[standard]
web3
numpy
pandas
scikit-learn
python-dotenv
httpx
pydantic
websockets
